Finish control signals + draft version menu

+++ ADC.h
@@ -4,6 +4,6 @@
 
 void ADC1_init(void) ;
 unsigned int read_ADC1 (void) ;
-void display_Measure(char* Measurement, char* Unit, int value);
+void display_Measure(char* Measurement, char* Unit, float value);
 
 #endif //__ADC_
